backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.5
SQLAlchemy==2.0.43
asyncpg==0.30.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
